Implemented reviewers suggestions radiobuttons id changed removed getDay() duplication fixed global feedback variable assertEquals replaced with assertThat
