+++ src/main/webapp/js/option.js
@@ -22,19 +22,6 @@ function onloadHandler() {
     getUserRole();
 }
 
-function getDate(date) {
-    var monthNames = [
-        "January", "February", "March",
-        "April", "May", "June", "July",
-        "August", "September", "October",
-        "November", "December"
-    ];
-    var day = date.getDate();
-    var monthIndex = date.getMonth();
-    var year = date.getFullYear();
-    return day + ' ' + monthNames[monthIndex] + ' ' + year;
-}
-
 function getUserRole() {
     console.log("role");
     fetch('/api/user/role', {
